Move utiltiy functions to utils.py and tests accordingly.
