+++ copulas/errors.py
@@ -1,0 +1,5 @@
+"""Copulas Exceptions."""
+
+
+class NotFittedError(Exception):
+    """NotFittedError class."""
