+++ copulas/multivariate/base.py
@@ -4,7 +4,8 @@ import pickle
 
 import numpy as np
 
-from copulas import NotFittedError, get_instance, validate_random_state
+from copulas.errors import NotFittedError
+from copulas.utils import get_instance, validate_random_state
 
 
 class Multivariate(object):
