+++ copulas/multivariate/gaussian.py
@@ -7,7 +7,9 @@ import numpy as np
 import pandas as pd
 from scipy import stats
 
-from copulas import (
+from copulas.multivariate.base import Multivariate
+from copulas.univariate import GaussianUnivariate, Univariate
+from copulas.utils import (
     EPSILON,
     check_valid_values,
     get_instance,
@@ -16,8 +18,6 @@ from copulas import (
     store_args,
     validate_random_state,
 )
-from copulas.multivariate.base import Multivariate
-from copulas.univariate import GaussianUnivariate, Univariate
 
 LOGGER = logging.getLogger(__name__)
 DEFAULT_DISTRIBUTION = Univariate
